# perf(atoms): Reduce size of `Atom`