**Description:**

This patch reduces the memory footprint

---

# ES parser comparison

## main

```

heap stats:    peak      total      freed    current       unit      count
  reserved:  512.1 MiB  512.1 MiB  256.0 MiB  256.1 MiB                        not all freed!
 committed:  512.1 MiB  512.1 MiB  256.0 MiB  256.1 MiB                        not all freed!
     reset:   13.2 MiB  775.2 MiB  896.9 MiB -121.7 MiB                        ok
   touched:      0          0       53.9 GiB  -53.9 GiB                        ok
  segments:     40        1.5 Ki     1.5 Ki       5                            not all freed!
-abandoned:      0          0          0          0                            ok
   -cached:      0          0          0          0                            ok
     pages:    2.0 Ki   956.6 Ki   956.5 Ki      28                            not all freed!
-abandoned:      0          0          0          0                            ok
 -extended:      0
 -noretire:      0
     mmaps:      0
   commits:      0
   threads:     10         10          0         10                            not all freed!
  searches:     0.0 avg
numa nodes:       1
   elapsed:      97.390 s
   process: user: 110.162 s, system: 0.846 s, faults: 0, rss: 351.0 MiB, commit: 512.1 MiB
```


## With this

```
heap stats:    peak      total      freed    current       unit      count
  reserved:  512.1 MiB  512.1 MiB  256.0 MiB  256.1 MiB                        not all freed!
 committed:  512.1 MiB  512.1 MiB  256.0 MiB  256.1 MiB                        not all freed!
     reset:   19.4 MiB  533.1 MiB  622.9 MiB  -89.7 MiB                        ok
   touched:      0          0       48.3 GiB  -48.3 GiB                        ok
  segments:     40        1.5 Ki     1.5 Ki       4                            not all freed!
-abandoned:      0          0          0          0                            ok
   -cached:      0          0          0          0                            ok
     pages:    1.9 Ki   852.7 Ki   852.7 Ki      28                            not all freed!
-abandoned:      0          0          0          0                            ok
 -extended:      0
 -noretire:      0
     mmaps:      0
   commits:      0
   threads:     10         10          0         10                            not all freed!
  searches:     0.0 avg
numa nodes:       1
   elapsed:      93.037 s
   process: user: 114.865 s, system: 1.098 s, faults: 0, rss: 348.0 MiB, commit: 512.1 MiB

```